Workaround a Qt bug causing crash in freebusymodel tests
QtWebEngineWidgets automatically creates a QOpenGLContext from a
a Q_COREAPP_STARTUP_FUNCTION which will cause a crash when we are
using only a QCoreApplication (so no OpenGL). This happens when the
library is loaded, so even if linked indirectly by another lib
(libkdepim in this case) so there's nothing we can do to prevent it
from being loaded.

Temporary workaround is to use QGuiApplication, which makes WebEngine
happy. This has been fixed in Qt 5.6.1 (QTBUG-51789)