Cache tags by name instead of gid
To set and retrieve category colors, a TagCache is used, where tags
are stored by gid. So far, the utf8 encoding of a tag's name was used
to retrieve a tag from the cache. In general, however, the gid of a tag
is not the same as its name. Since the gid may be unknown when looking
for a tag's color, this change switches to the tags' names for keys.

BUG: 333754
FIXED-IN: 16.08.1

Differential Revision: https://phabricator.kde.org/D2434