Improve the small calendars in the printed headers
The small monthly calendars in the headers of printed pages squish the
rows of dates together, and cut off the bottoms of the month names.
`CalPrintPluginBase::drawSmallMonth()` calculates the row height in
pixels, then fudges that number for use as a font size in points,
unsuccessfully.

This patch sets the font size in pixels, and uses `QFontMetrics::height()`
to leave room for month names with letter glyphs that have descenders.